requirements.txt
flask-cors==2.1.2
slumber==0.7.1
pytest==2.9.0
pytest-flask==0.10.0
Flask-Cache==0.13.1
SQLAlchemy==1.1.1
flask-sqlalchemy==2.1
sqlalchemy-utils==0.32.9
python-decouple==3.0
flask-admin==1.4.2
itsdangerous==0.24
flask-validator==1.2.1
flask-fixtures==0.3.7
flask-restful==0.3.5
coverage==4.2
pytest-cov==2.4.0
flask_restplus==0.9.2
coveralls==1.1
Flask-Mail==0.9.1
